fix tests on macOS
The tests assume that the config dir is ~/.config and the data dir is
~/.local/share. Neither is true on macOS, and this causes several tests
to fail:

	FAILED tests/test_emborg.py::test_emborg_without_configs[4] - AssertionError: Test ‘labor’ fails.
	FAILED tests/test_emborg.py::test_emborg_with_configs[5] - AssertionError: Test ‘periphery’ fails.
	FAILED tests/test_emborg.py::test_emborg_with_configs[25] - AssertionError: Test ‘build’ fails.
	FAILED tests/test_emborg.py::test_emborg_overdue[1] - AssertionError: Test ‘predator’ fails.
	FAILED tests/test_emborg.py::test_emborg_overdue[2] - AssertionError: Test ‘smelt’ fails.

Rather than changing the tests to dynamically use different test paths
based on the OS[1], it's simpler to change emborg's behavior at test
time to always use the Linux paths. That is what this change does.

Also adds tests/Users to .gitignore, which is the macOS equivalent of
tests/home.

[1] This was in fact the first thing I tried. But as I got into it the
    added complexity didn't seem justified.